wgpu: Directly copy BitmapData gpu buffer to Stage3D texture
We now immediately flush our in-progress buffer CommandEncoder
after executing a `CopyBitmapToTexture` command, to ensure that
we don't pick up any future avm2-side changes to the BitmapData
gpu texture.

The Vec -> gpu buffer copy implementation is moved to a new
`CopyBytesToTexture` command, which we use for copying AGAL
textures.