web: Display loading screen while resolving the SWF URL
Previously, when entering a URL in the extension player, it had been
resolved before the loading screen had been displayed. As resolving can
include testing for server availability (to decide whether HTTPS or HTTP
is used), this could create a wait time without the loading screen.
This has been fixed. The loading screen is now displayed while the URL
is resolved. Additionally, the timeouts have been increased (they were
low because of the lack of a loading animation).

A bug that caused the document title to not be set correctly if the SWF
URL was the main page of a domain has also been fixed.