desktop: Improve, update and correct Windows metadata
The metadata packaged in the Windows executable has been improved,
updated and corrected.

Previously, a static RC file had been used to provide Windows metadata.
However, it had been outdated, wrong (the InternalName value was not in
line with the Microsoft documentation) and lacked information (e.g. the
debug flag had never been set).

Additionally, a static RC file is not a good way to package Windows
metadata as it's hard to dynamically update. Therefore, it has been
replaced by the winresource crate.
The winresource crate is now used by the build script to dynamically
generate an up-to-date RC file and embed it in the Windows executable.