desktop: Add DMG creation to the macOS build workflow
The creation of a DMG file has been added to the macOS build workflow.
It replaces the tar.gz file for macOS and is created and uploaded for
each release. dmgbuild is used to create it.
A background image for the DMG file has been added as well.

Additionally, some wording has been improved.