web: Fix error messages outside the extension player
The error messages outside the extension player have been fixed.
Ruffle now correctly determines whether the last load command originated
from the extension player and displays the error messages specific to it
(invalid URL / local URL / URL with unsupported protocol entered) only
in that case. The advice to try entering http explicitly has also been
removed if the URL is not loaded from the extension player.

Additionally, if the website now embeds a URL with an unsupported
protocol (e.g. file), a new specific error message about the website
being misconfigured is now displayed instead of a button to open the URL
in the extension player (which itself will result in an error).