desktop: Add AppImage creation to the Linux build workflow
The creation of an AppImage has been added to the Linux build workflow.
It is now bundled inside the Linux tar.gz instead of the raw binary.
This leads to a higher distribution compatibility and a better desktop
integration.
Linuxdeploy is used to create it.

The Linux build workflow is now run inside a docker container with the
oldest still-supported LTS version. This should be done when building
AppImages to be compatible with all still-supported Linux distribution
versions. In order to use the container, a new step has been added which
installs the required software the default GitHub runner already had
preinstalled.

Additionally, an SVG Ruffle icon has been added, as well as a 512x512px
raster version. They have been created out of the existing macOS SVG
icon (which has been renamed). They are used for the new AppImage.
The desktop entry file has also been updated to use the correct icon and
executable name.