desktop: Add single arch. package creation to the macOS build workflow
The creation of single architecture packages has been added to the macOS
build workflow. It now produces one x86-64, one arm64 and one universal
package.
As the process of packaging the executable into an app bundle inside a
DMG image is the same for all packages, the common logic has been
extracted and moved into a new shell script. This shell script is called
in both jobs with the necessary parameters and environment variables.
This also allows for the arm64 image to be packaged using a newer
compression format and filesystem, as older systems don't have to be
supported.

In order for the single architecture packages to be built, the Safari
build artifact needs to be downloaded. Therefore, the build job now
depends on the build-web job, and the necessary artifacts are downloaded
in it.