web: Implement URL rewrite rules
These rules are configurable and they allow rewriting URLs used in SWFs.
The main rationale behind them is to rewrite obsolete links to
working ones in order to preserve old content.

In order to add a rewrite rule, add a `urlRewriteRules` config option
which is an array of rewrite rules.  Each rule consists of a matcher
(either a string or a regex), and a replacement, e.g.:

    window.RufflePlayer.config = {
        "urlRewriteRules": [
            // Rewrites to a relative URL
            [/^https?:\/\/example.com\/(.*)\.swf$/, "$1.swf"],
            // Rewrites to an absolute URL
            [/^(.*)\.test\.swf$/, "$1.swf"],
            // Rewrites using a string (exact match, not a regex)
            ["http://old.example.com/site.html", "http://new.example.com/other_site.html"],
        ],
    };