docs: Correct Apache Flex SDK compiling guide
Change -compiler.debug to -debug to make it work in powershell.
It gets tripped up by the period in the argument name.

Change -output to -o to make it shorter.

Make note of the capitalization of the file name.
It won't compile unless the casing matches that of the class name's.