desktop: Add Alt+Enter back as a hidden shortcut on Windows
Previously Alt+Enter was a hidden shortcut that was used to toggle
full screen, and it is used by Flash Player to toggle full screen
on Windows too.

As Windows users sometimes expect this shortcut to work, this
patch brings it back as a hidden shortcut (until we add some kind of preferences).