web: Don't set the clipboard when notified about clipboard change
This patch introduces WebUiBackend::set_clipboard_content_buffer which
is able to set the clipboard buffer without also setting the clipboard.
This method is then used in places where Ruffle is notified about a paste
event, so that it doesn't have to set the clipboard to the same value.

This fixes a borrow checker error during pasting.