desktop: Synchronize window theme on Linux
Set window decorations theme so that they are always
synchronized with the GUI theme on Linux.
Note that this does not apply to other OSes,
as we rely on window theme changing automatically there.

This fixes window decorations desync on Linux Wayland, which is caused
by sctk-adwaita not watching on system theme changes.