text: Calculate text width dynamically for autosize
When horizontal autosize is effective, text width is calculated
dynamically and is equal to the width of the longest line in text.

That's why this patch makes the text width optional, and in case
it's None, the text is laid out two times—first time in order to calculate
the proper text width, and the second time knowing what width should be used.