text: Fix EditText::maxhscroll
This patch fixes EditText::maxhscroll:

1. the gutter is taken properly into account,
2. text fields with a width smaller than 4px are now supported,
3. input fields are now properly supported.

Input fields allocate a bit more space for horizontal scroll.
It's exactly equal to W / 4, where W is the width of the window
where the text is visible (bounds without the gutter).
This formula has been derived experimentally using a linear regression
p ~ a * q + b, where p is the field's width, and q is the difference
between maxScrollH of input and non-input fields with the width.