core: Implement tab_skip compatibility flag
This flag configures the behavior of skipping some objects when tabbing.
Flash Player skips those objects by default, however it's possible that
some content is not aware of this (e.g. when objects are placed on a grid),
and it results in reduced accessibility, as users cannot tab some objects.