render: Flush draws before updating bitmap texture
This fixes an issue where new draws were drawn over an old GPU buffer.
For instance, consider the following scenario:

    1. draw (GPU)
    2. full fill (CPU)
    3. draw (GPU)

When full fill is executed, the GPU buffer is discarded without a sync,
but we have to make sure we wait for any queued draws in order to
truly discard 1.

Co-authored-by: Nathan Adams <dinnerbone@dinnerbone.com>