avm1: Do not unwrap parent when removing objects pending removal
Turns out that sometimes objects pending removal do not have a parent.

I wasn't able to reproduce it, but it happens as we can see in error
reports.

Semantically this should be fine, as we mainly operate on the parent
here, but it's hard to be sure considering it's not reproducible.
It's better than panicking for sure.