exporter: add a frame cache
This commit adds a frame cache which limits the amount of frames that
can be rendered into memory before writing to disk and getting more
frames. This can save memory on big 4k frames, especially if you have
alot of them as it can fill up 32 gigs of ram pretty fast.