web: Fix context menu positioning on RTL pages
This patch fixes a bug where the context menu was positioned incorrectly
in RTL contexts. The bug was caused by the fact that x,y coordinates in
JS events are not RTL aware, whereas the `transform` CSS property is.

This patch changes CSS transform usage to top/left/right absolute
positioning.

Additionally, it adds support for RTL context menus, but sets their
direction permanently to LTR, because we don't have an easy way of
detecting LTR/RTL browsers. (Note that the context menu should not
follow the direction of the page, but the direction of the browser.)