debug_ui: Add convenience action to open root movie clip (if it exists)
It's annoying from a UX standpoint that one has to go through
View Stage Info -> Children -> (Root MovieClip) -> MovieClip (tab)
to access the "main" frame list, in order to get to the interesting
parts of a movie (for debugging or otherwise).

This commit adds an action that directly opens the MovieClip tab
of the root MovieClip, if it exists.