desktop: Load fallback UI fonts properly
Before this patch, we were loading only one font, based on the OS and
language selected. This approach was not perfect, as the UI can have
characters from multiple fonts, irrespectively of selected language.

This patch changes the way fonts are loaded, and provides multiple
fallback fonts for different categories of characters (Korean, Chinese,
Japanese, etc.).

Fonts are ordered by their priority which depends on the language.
That is because some languages (Japanese, Chinese) have overlapping
characters.