debug_ui/display_object: Remove collapsing header for frame list
Instead, we embed it directly in a compact way.

My guess is that the original reason it was put into a CollapsingHeader
is that the original frame list tanked the performance when shown, but
this has been fixed by using `TableBody::rows` to only render the visible
rows.

This allows quick access to the frame list, which is helpful for quickly
seeking to the parts of a movie we're interested in.