core: Clamp min coordinates in color_transform
The current behavior assumes the max value can be larger than the
target size, but that the min can't.

This reasoning seems faulty, since one could pass (min:100,max:150)
to a target of size 50, in which case both min and max are larger,
but only max would get clamped.

We now clamp both min and max to correctly enforce invariants.