desktop: Enable fontconfig feature by default
Fontconfig is a library available for most OSes, but most commonly used
in Linux distributions. Windows and MacOS provide their own mechanisms
of looking up fonts.

Even if this feature is unlikely to be useful for Windows and MacOS,
enabling it doesn't introduce bloat, as it only contains bindings
for a shared library.