tests: Add support for filters in image checks
Filters are useful when e.g. the given check only applies to a specific
platform. It's very common that a test output looks different on
different platforms. With filters we can write specific checks for
specific platforms instead of increasing tolerance for all of them.

An example configuration might look like this:

    [[image_comparisons.output.checks]]
    filter = 'not(arch = "aarch64")'
    tolerance = 2

    [[image_comparisons.output.checks]]
    filter = 'arch = "aarch64"'
    tolerance = 2
    max_outliers = 10

In this configuration, we allow 10 outliers for aarch64, while not
allowing any for all other architectures.