avm1: Fix timeline scope on target clip removal (close #9863)
When an activation's target clip is removed, fall back to the base
clip, or invalidate it if the base clip is also removed. Also set
the timeline scope to match the new target clip, using the root
clip if it's invalid. Additionally, don't continue script execution
on an `action_call` when the base clip is removed.