core: Allow loading fonts into memory lazily
This patch improves memory footprint by preventing fonts from being
loaded eagerly into memory. Instead, it's possible to pass a file
reference to a font, which will be used to load data from the font file
when needed. This mechanism is used on desktop.

This patch also introduces a preloading mechanism, that makes sure we
read from the font file once per a text fragment, and not for each
glyph.