avm2: Add a buggy SWF<11 integer overflow in sort comparison
Flash Player had a bug in SWF<11 where the numeric comparison function
overflowed for certain numbers.  This patch replicates this behavior.

Currently it only works for Flash Player 32-bit, as Ruffle does not
support the 64-bit runtime.  On 64-bit the behavior differs a bit due to
different integer representations.

The test `from_avmplus/as3/Array/regress_524122` is written in such
a way that it passes for both 32- and 64-bit behaviors.