text: Improve rendering underlines
This patch improves rendering underlines, specifically:
* They are rendered as primitives instead of drawings.
* They are being rendered on the proper height now.
* They scale properly (are always 1px thick).

The ability to render drawings is left as-is for now, as rendering
images within HTML is architecturally very similar to rendering those
drawings.  In order to preserve code structure that will be useful when
implementing images in HTML, the drawing layout content is left and
marked as unused for now.