tests: Disable default font feature by default
Having default font enabled makes it easier to write a bad test, because
it's a Ruffle-specific behavior that Flash doesn't have.  It means that
every test that uses the default font (i.e. changes output with default
font enabled) is probably not based on FP's output, but on Ruffle's
output.

This PR adds a property `with_defaut_font`, which, when set, will load
the default font and the test will work as before.  It also marks all
tests that rely on this behavior.  Almost all of them should be fixed.