core: Make code more readable for PixelRegion
The previous code did make sure that x_max and y_max could not be less then x_min and y_min. But this is already ensured because saturating_add won't wrap around and the arguments to the function (width, height) will never be negative since they're unsigned.

This commit removes the extra checks while keeping the functionality equal.