ci: Calculate delta coverage using diff-cover
Now each coverage report also includes delta coverage, i.e. coverage
scoped to current changes.  For PRs, it's code modified in the PR,
for branches it's code since the last push.

The delta coverage report is included as a separate report and also
a markdown summary is posted on GitHub Actions.