render: Support transparency in masks
When CAB is enabled on both mask and maskee, masks are rendered
differently; instead of masking pixels based on hit test, pixels are
masked based on the value of alpha in the mask. This is more akin to
BlendMode.ALPHA.

This patch adds support for alpha masks (or CAB masks), which are
rendered using a shader, compared to a stencil used by regular masks.