avm2: Refactor Array.splice to ArrayStorage
This refactor moves splice implementation from AVM2 to ArrayStorage.
The goal was to improve performance of splice, which was only partially
successful, as it works better, but the bottleneck is still the weird FP
behavior with holes and the prototype.

Nonetheless, this looks like a step in the right direction.