core: Fix drawing Stage into a bitmap
This patch moves `Stage.render`'s logic into `Stage.render_viewport`.
The reason for that is you can render Stage as any other display object,
and rendering the viewport in `render` results in a completely different
behavior for Stage.

This fixes an issue where rendering Stage into a bitmap resulted in
the viewport being rendered.