web: Improve asserting on traces in tests
Before this patch there could be potential race conditions on traces if
they were produced asynchronously by the SWF. The old logic waited for
any trace and then asserted on all available traces, which meant that
we could include some unintended traces or assert before all traces have
been produced.

The new approach ensures that we wait for as many traces as we need.