avm2: Add Value normalization
This patch adds Value::normalize, which normalizes the given Value to
an equivalent one.

It should be fine to call this method whenever, it does not change
semantics, but has an effect on performance only.

Flash Player does this normalization on every atom instantiation,
but for Ruffle it's too inefficient (we aren't doing any allocs).
However, there are some observable behaviors that result from it, and
that's why this method is provided in order to cover such cases.