avm2: Allow any i32 in Value::Integer
This is a performance optimization, as it gets rid of range checks
when operating on values. Since we already implement value
normalization, we don't have to make sure the value is the proper type
at all times, only when differentiating between Number and Integer.
This allows us to drop range checks and allow any Integer and any Number
even if they are not representable in Flash Player.

This should improve performance in most cases, where we don't need a
range check. This introduces overflow checks in op_add and op_subtract,
but they should be faster than range checks which happened after each op
anyway.

When overflowing, the values are added/subtracted as i64 and only then
cast to f64, as that produced better results in benchmarks.