web: Fix flakiness of mouse_wheel_avm2
wdio was randomly causing the browser to fire pointerenter and
pointerleave events, which resulted in the test being flaky.
When pointerleave event was sent before moving the mouse, there was no
hovered object and scroll was not dispatched to the object it should
have been.

This patch uses wdio's .moveTo() to update the hover before dispatching
scroll event which fixes this issue.