core: Use __jcvt() intrinsic on AArch64 for f64→i32 conversion
In #21780, an optimisation has been added to use the fjcvtzs ARMv8.3
instruction when available, to convert a f64 into an i32.  This made me
wonder why core::arch::aarch64 didn’t have an intrinsic for this
instruction, so I implemented it in stdarch[1], which got pulled in
Rust[2] (see the tracking issue[3]) and then stabilized in 1.95.0.

This PR makes use of this new intrinsic to remove the unsafe asm!()
block, and simplify the code.

[1] https://github.com/rust-lang/stdarch/pull/1938
[2] https://github.com/rust-lang/rust/pull/148402
[3] https://github.com/rust-lang/rust/issues/147555