core: Add compatibility rule to rewrite fpdownload.adobe.com
This domain hosts SWZ files and does not work on web due to CORS.
This fixes content that requires those SWZ files and the website owner
does not provide them. As a bonus we don't have to worry about
reliability of fpdownload.adobe.com and the fact that some day it will
stop working.