desktop: Implement FreeType device font renderer
It's our second font renderer, this time for desktop, which allows using
the OS glyph rendering capabilities for displaying fonts in Ruffle.

It doesn't impact font selection (we still use fontconfig for that),
only font rendering (rasterization). It's supported only for Linux at
this moment (similarly to fontconfig). In theory it's possible to use
FreeType on other OSes, but people don't have that installed. We can
think about bundling FreeType for other OSes in the future.

Why am I doing this? Two reasons.

1. Fonts will look nicer on desktop (in the future, currently they look
   terrible). FreeType can render fonts at different resolutions and can
   provide high quality fonts even when their height is small. The
   embedded renderer renders fonts just like any other shape, which is
   not legible.

2. This will FINALLY allow testing font renderers. Previously we only
   had the web font renderer, and testing it ranged from difficult to
   impossible. Now we can easily test FreeType renderer and the whole
   bitmap glyphs machinery (which is also used on web).

Currently the renderer does the same hack as the web renderer, which
renders glyphs at a static resolution and scales bitmaps. This causes
fonts to look absolutely terrible, but having tests will allow to change
that in the near future.