avm1: More accurate semantics for `instanceof` and `CastOp`
- only coerce RHS when the LHS is object-like
- `prototype` access should ignore __proto__, properties, and version attributes

This exposes some Flash weirdness; in FP, the following code traces `false` if
placed on frame 1, but hangs if placed on any other frame.

```
a = {};
b = {};
a.__proto__ = b;
b.__proto__ = a;
trace(a instanceof b);
```

Prior to this PR, Ruffle raised an `Error::PrototypeRecursionLimit`, for slightly
wrong reasons (we tried to crawl b's `__proto__` chain to find a `prototype`
property); now it returns `false`.

This affects Gnash's `Inheritance-*` tests, as they run the above code on frame 2,
and so hang in FP.