avm1: Add _global.textRenderer
In addition to `_global.flash.text.TextRenderer`, there's also
`_global.textRenderer` which is the same thing.

AVM1 globals assigns objects to the variable `o` in order to set
properties without repeating the full path. It looks like `TextRenderer`
was assigned to the variable `textRenderer` instead by accident, causing
it being defined in globals.