desktop: Simplify value calculation in PathOrUrlField
Before this refactor, the URL was always inferred from its string
representation.  After this refactor, both have to be inferred together
for each change, which means we have better control over the value.
It means that we can encode information in the URL that is not present
in the representation that comes e.g. from the picker.